Remove a bunch of components from pages folder
