+++ frontend/sheet-manager/pages/characters/new/index.jsx
@@ -1,13 +1,13 @@
 import { useContext, useState } from 'react'
 import ScreenSlideProvider, { ScreenSlideContext } from '@/src/contexts/screen_slide_context'
-import Role from './steps/role'
-import Specie from './steps/specie'
-import Culture from './steps/culture'
-import Grimo from './steps/grimo'
-import Attributes from './steps/attributes'
-import StarterWeapon from './steps/starter_weapon'
-import Spells from './steps/spells'
-import Details from './steps/details'
+import Role from '@/src/components/characters/steps/role'
+import Specie from '@/src/components/characters/steps/specie'
+import Culture from '@/src/components/characters/steps/culture'
+import Grimo from '@/src/components/characters/steps/grimo'
+import Attributes from '@/src/components/characters/steps/attributes'
+import StarterWeapon from '@/src/components/characters/steps/starter_weapon'
+import Spells from '@/src/components/characters/steps/spells'
+import Details from '@/src/components/characters/steps/details'
 import Overview from './overview'
 import { AnimatePresence } from 'framer-motion'
 import useLocalStorageState from '@/src/utilitaryHooks/use_local_storage_state'
