+++ frontend/sheet-manager/pages/login/index.jsx
@@ -1,7 +1,7 @@
-import Button from '../components/button'
-import Input from '../components/input'
-import Logo from '../components/logo'
-import AppLink from '../components/app_link'
+import Button from '@/src/components/button'
+import Input from '@/src/components/input'
+import Logo from '@/src/components/logo'
+import AppLink from '@/src/components/app_link'
 import { useEffect, useState } from 'react'
 import Router from 'next/router'
 import useLogin from '../../hooks/use_login'
