+++ frontend/sheet-manager/pages/signup/index.jsx
@@ -1,11 +1,11 @@
-import Logo from '../components/logo'
-import Button from '../components/button'
-import Input from '../components/input'
-import Container from '../components/container'
+import Logo from '@/src/components/logo'
+import Button from '@/src/components/button'
+import Input from '@/src/components/input'
+import Container from '@/src/components/container'
 import { BsArrowLeftCircle } from 'react-icons/bs'
-import Checkbox from '../components/checkbox'
-import AppLink from '../components/app_link'
-import Title from '../components/title'
+import Checkbox from '@/src/components/checkbox'
+import AppLink from '@/src/components/app_link'
+import Title from '@/src/components/title'
 import { toast } from 'react-toastify'
 import { useState } from 'react'
 import Router from 'next/router'
