+++ frontend/sheet-manager/src/components/characters/_navigation.jsx
