+++ frontend/sheet-manager/src/components/logo.jsx
