+++ frontend/sheet-manager/src/components/waves.jsx
